# docs/openapi.yaml
openapi: 3.1.0
info:
  title: QK 教培 CRM 管理系统 API
  version: 1.1.12
  summary: 线索、商机、客户全流程管理接口
  description: |
    面向教培行业销售团队的后台管理接口，覆盖线索录入、分配、跟进、转商机、再分配、
    再跟进、转客户的全流程，以及部门、角色、用户、课程、活动等基础数据维护。

    ## 通用约定

    ### 统一响应结构
    所有接口返回 `Result` 对象：`code` 为 `1` 表示成功、`0` 表示失败。
    **业务失败同样返回 HTTP 200**，由调用方依据 `code` 判断；`msg` 为可直接展示给用户的中文提示。
    无业务数据时 `data` 为 `null`，且该字段会**整个省略**（服务端 `Result` 标注 `@JsonInclude(NON_NULL)`），
    判断成功请用 `code == 1`，不要依赖 `data` 字段是否存在。

    ### 认证
    除 `POST /login` 外，所有接口都要求在请求头 `token` 中携带登录接口签发的 JWT 令牌。
    令牌缺失、被篡改、已过期，或令牌对应的账号已被停用 / 删除时，接口直接返回 **HTTP 401 且响应体为空**，
    不返回 `Result` 结构。
    令牌默认有效期 24 小时。

    ### 授权（权限点模型）
    已登录不等于有权限。每个接口在代码里声明自己需要的**权限点**（如 `clue:track`、`user:delete`，
    完整目录见 `GET /permissions`），账号能不能调用取决于「它的角色被授予了哪些权限点」。
    角色是数据，可以自由增删改，**改角色名/标识不影响任何权限**；授权关系存在 `role_permission` 表，
    用 `GET /roles/{id}/permissions` 查看、`PUT /roles/{id}/permissions` 覆盖式配置（改完立即生效）。
    未声明权限的接口对所有人拒绝（默认拒绝），唯一例外是 `POST /login`。
    内置超级管理员角色（`role.is_super = 1`）天然拥有全部权限，不可删除、也不支持单独调整权限。
    登录响应里的 `permissions` 即当前账号的权限码列表，前端用它控制菜单与按钮（不要用 `roleLabel`）。

    ### 分页
    分页接口的查询参数为 `page`（页码，从 1 开始，默认 1）与 `pageSize`（每页条数，默认 10，**单页最大 200**），
    响应数据统一为 `{ "total": 总记录数, "rows": [...] }`。
    `page` / `pageSize` 必须为正整数：传空值、0、负数或超过 200 时返回 `code = 0`，`msg` 为对应的字段提示。

    ### 错误语义（HTTP 状态码）
    | 状态码 | 触发场景 | 响应体 |
    | --- | --- | --- |
    | `200` | 成功；以及**业务失败**（数据不存在、唯一键冲突、文件类型不合法、状态不允许等） | `Result`，失败时 `code = 0`，按 `msg` 提示 |
...

...

...

...

...

...

...

...
servers:
...
security:
  - token: []
...
paths:
...
  /courses:
    get:
      tags: [课程管理]
      summary: 条件分页查询课程
      operationId: listCourses
      parameters:
        - name: name
          in: query
          description: 课程名称，模糊匹配
          schema: { type: string }
        - name: subject
          in: query
          description: 学科
          schema: { $ref: '#/components/schemas/Subject' }
        - name: target
          in: query
          description: 适用人群
          schema: { $ref: '#/components/schemas/CourseTarget' }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: 分页结果，按更新时间倒序
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultCoursePage'
    post:
      tags: [课程管理]
      summary: 新增课程
      description: 该接口会写入一条操作日志。
      operationId: addCourse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseBody'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          $ref: '#/components/responses/Ok'
    put:
      tags: [课程管理]
      summary: 修改课程
      description: 该接口会写入一条操作日志。
      operationId: updateCourse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseBody'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          $ref: '#/components/responses/Ok'
  /courses/list:
    get:
      tags: [课程管理]
      summary: 查询所有课程（下拉框）
      operationId: listAllCourses
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: 全量课程列表，按ID升序
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultCourseList'
  /courses/subject/{subject}:
    get:
      tags: [课程管理]
      summary: 根据学科查询课程
      operationId: listCoursesBySubject
      parameters:
        - name: subject
          in: path
          required: true
          description: 学科
          schema: { $ref: '#/components/schemas/Subject' }
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: 课程列表，无匹配时返回空数组
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultCourseList'
  /courses/{id}:
    get:
      tags: [课程管理]
      summary: 根据ID查询课程
      operationId: getCourseById
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: 课程信息；不存在时 `code = 0`、`msg = 课程不存在`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultCourse'
    delete:
      tags: [课程管理]
      summary: 删除课程
      description: 该接口会写入一条操作日志。
      operationId: deleteCourse
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          $ref: '#/components/responses/Ok'
...
  /clues:
    get:
      tags: [线索管理]
      summary: 线索列表查询
      description: |
        默认只返回仍在跟进的线索：`status` 为 `4 伪线索` 与 `5 转为商机` 的线索不出现在此列表。
        显式传 `status` 时按其取值筛选（前端状态下拉里包含这两个状态，选中后应能看到对应数据）。
        按修改时间倒序。
      operationId: listClues
      parameters:
        - name: clueId
          in: query
          description: 线索ID，精确匹配
          schema: { type: integer, format: int64 }
        - name: phone
          in: query
          description: 手机号，模糊匹配
          schema: { type: string }
        - name: status
          in: query
          description: 线索状态；不传时排除 4 伪线索、5 转为商机
          schema: { $ref: '#/components/schemas/ClueStatus' }
        - name: channel
          in: query
          description: 线索来源
          schema: { $ref: '#/components/schemas/Channel' }
        - name: assignName
          in: query
          description: 归属人姓名，模糊匹配
          schema: { type: string }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: 分页结果，`rows` 中 `trackRecords` 与 `record` 为 `null`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultCluePage'
    post:
      tags: [线索管理]
      summary: 新增线索
      description: |
        新增的线索一律置为 `1 待分配` 且归属人清空，请求体中的 `status`、`userId` 不生效。
        手机号唯一，重复时返回 `msg = 该手机号已录入线索`。该接口会写入一条操作日志。
      operationId: addClue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClueBody'
            example:
              phone: '15100000001'
              channel: 1
              name: 卫丹
              gender: 2
              age: 24
              wechat: wxweidan
              qq: '8450313640'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          $ref: '#/components/responses/Ok'
    put:
      tags: [线索管理]
      summary: 跟进线索
      description: |
        将线索状态置为 `3 跟进中`，并新增一条 `type = 1` 的正常跟进记录。
        请求体中除 `id` 外的字段为可选，未传字段保持原值；`record` 为本次跟进内容，写入跟进记录。
        该接口会写入一条操作日志。
      operationId: trackClue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClueTrackBody'
            example:
              id: 55
              subject: 1
              level: 3
              nextTime: '2025-05-22 12:00:00'
              record: 了解一下目前 java 方向的就业前景
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          $ref: '#/components/responses/Ok'
  /clues/pool:
    get:
      tags: [线索管理]
      summary: 线索池列表查询
      description: 只返回伪线索（`status = 4`），供管理员重新分配；并关联带出活动名称。按修改时间倒序。
      operationId: listPoolClues
      parameters:
        - name: clueId
          in: query
          description: 线索ID，精确匹配
          schema: { type: integer, format: int64 }
        - name: phone
          in: query
          description: 手机号，模糊匹配
          schema: { type: string }
        - name: channel
          in: query
          description: 线索来源
          schema: { $ref: '#/components/schemas/Channel' }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: 分页结果，`rows` 中 `activityName` 为活动名称
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultCluePage'
  /clues/assign/{clueId}/{userId}:
    put:
      tags: [线索管理]
      summary: 分配线索给指定用户
      description: 将线索归属人设为指定用户并置为 `2 待跟进`。该接口会写入一条操作日志。
      operationId: assignClue
      parameters:
        - name: clueId
          in: path
          required: true
          description: 线索ID
          schema: { type: integer, format: int64 }
        - name: userId
          in: path
          required: true
          description: 被分配的用户ID
          schema: { type: integer, format: int64 }
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          $ref: '#/components/responses/Ok'
  /clues/false/{id}:
    put:
      tags: [线索管理]
      summary: 将线索标记为伪线索
      description: |
        将线索置为 `4 伪线索`，并新增一条 `type = 0` 的跟进记录，记录伪线索原因与备注。
        该接口会写入一条操作日志。
      operationId: markFalseClue
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarkFalseClueBody'
            example:
              reason: 2
              remark: 电话停机了
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          $ref: '#/components/responses/Ok'
  /clues/toBusiness/{id}:
    put:
      tags: [线索管理]
      summary: 将线索转为商机
      description: |
        将线索置为 `5 转为商机`，并按线索信息复制生成一条商机：
        商机重新走分配流程，因此归属人与下次跟进时间被清空，状态为 `1 待分配`，同时记录来源线索ID。
        该接口会写入一条操作日志。
      operationId: convertClueToBusiness
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          $ref: '#/components/responses/Ok'
  /clues/{id}:
    get:
      tags: [线索管理]
      summary: 根据ID查询线索详细信息
      operationId: getClueById
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: |
            线索详细信息，`trackRecords` 为该线索的全部跟进记录（按创建时间倒序），
            每条记录带出跟进人姓名。线索不存在时 `code = 0`、`msg = 线索不存在`。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultClue'
...
  /customers:
    get:
      tags: [客户管理]
      summary: 客户列表查询
      operationId: listCustomers
      parameters:
        - name: phone
          in: query
          description: 手机号，模糊匹配
          schema: { type: string }
        - name: name
          in: query
          description: 客户姓名，模糊匹配
          schema: { type: string }
        - name: channel
          in: query
          description: 渠道来源
          schema: { $ref: '#/components/schemas/Channel' }
        - name: subject
          in: query
          description: 意向学科
          schema: { $ref: '#/components/schemas/Subject' }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: 分页结果，按创建时间倒序，`rows` 中 `courseName` 为意向课程名称
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultCustomerPage'
    post:
      tags: [客户管理]
      summary: 新增客户
      description: |
        手工新增的客户没有来源商机，`businessId` 为空。
        手机号唯一，重复时返回 `msg = 该手机号已录入客户`。该接口会写入一条操作日志。
      operationId: addCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerBody'
            example:
              phone: '15600000001'
              channel: 1
              name: 库家明
              gender: 1
              age: 22
              wechat: kujiaming
              qq: '3353439142'
              degree: 4
              jobStatus: 1
              subject: 1
              courseId: 1
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          $ref: '#/components/responses/Ok'
    put:
      tags: [客户管理]
      summary: 修改客户
      description: 请求体中的 `businessId` 会被服务端忽略并置空，不允许通过修改接口篡改来源商机。该接口会写入一条操作日志。
      operationId: updateCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerBody'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          $ref: '#/components/responses/Ok'
  /customers/{id}:
    get:
      tags: [客户管理]
      summary: 根据ID查询客户
      operationId: getCustomerById
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: 客户信息；不存在时 `code = 0`、`msg = 客户不存在`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultCustomer'
...
components:
  securitySchemes:
    token:
      type: apiKey
      in: header
      name: token
      description: |
        登录接口返回的 JWT 令牌。缺失、被篡改或已过期时接口返回 HTTP 401 且响应体为空。
  parameters:
    Id:
      name: id
      in: path
      required: true
      description: 主键ID，64 位整型（对应数据库 bigint unsigned）
      schema: { type: integer, format: int64 }
...
    Page:
      name: page
      in: query
      description: 页码，从 1 开始
      schema: { type: integer, format: int32, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      description: 每页记录数
      schema: { type: integer, format: int32, minimum: 1, default: 10 }
  responses:
    Ok:
      description: 操作成功，`code = 1`、`data` 为 `null`
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Result'
    Unauthorized:
      description: 令牌缺失、被篡改或已过期。响应体为空，不返回 `Result` 结构。
...
  schemas:
...
    Result:
      type: object
      description: 统一响应结构。业务失败同样返回 HTTP 200，由 `code` 区分。
      properties:
        code:
          type: integer
          description: 1:成功, 0:失败
          enum: [1, 0]
        msg:
          type: string
          description: 提示信息，成功时固定为 `success`
        data:
          description: 业务数据，无数据时为 `null`
      required: [code, msg]
    DateTime:
      type: string
      description: 时间，输出格式 `yyyy-MM-dd HH:mm:ss`
      example: '2025-06-01 00:00:00'

    # ---------- 枚举 ----------
    Channel:
      type: integer
      format: int32
      description: 渠道来源，1:线上活动, 2:推广介绍
      enum: [1, 2]
    Subject:
      type: integer
      format: int32
      description: 意向学科，1:AI智能应用开发(Java), 2:AI大模型开发(Python), 3:AI鸿蒙开发, 4:AI大数据, 5:AI嵌入式, 6:AI测试, 7:AI运维
      enum: [1, 2, 3, 4, 5, 6, 7]
    Gender:
      type: integer
      format: int32
      description: 性别，1:男, 2:女
      enum: [1, 2]
    Degree:
      type: integer
      format: int32
      description: 学历，1:高中, 2:中专, 3:大专, 4:本科, 5:硕士, 6:博士, 7:其他
      enum: [1, 2, 3, 4, 5, 6, 7]
    JobStatus:
      type: integer
      format: int32
      description: 在职情况，1:在职, 0:离职
      enum: [0, 1]
    ClueStatus:
      type: integer
      format: int32
      description: 线索状态，1:待分配, 2:待跟进, 3:跟进中, 4:伪线索, 5:转为商机
      enum: [1, 2, 3, 4, 5]
    ClueLevel:
      type: integer
      format: int32
      description: 意向等级，1:近期学习, 2:打算学习(考虑中), 3:进行了解, 4:打酱油
      enum: [1, 2, 3, 4]
...
    FalseReason:
      type: integer
      format: int32
      description: 伪线索原因，1:空号, 2:停机, 3:竞品, 4:无法联系, 5:其他
      enum: [1, 2, 3, 4, 5]
...
    CourseTarget:
      type: integer
      format: int32
      description: 适用人群，1:小白学员, 2:中级程序员, 3:初级程序员（页面原型比接口文档多出「初级程序员」一档，故放开为 1~3 的超集）
      enum: [1, 2, 3]
...
    CourseVO:
      type: object
      properties:
        id: { type: integer, format: int64, description: 课程ID }
        subject: { $ref: '#/components/schemas/Subject' }
        name: { type: string, description: 课程名称 }
        price: { type: integer, format: int32, description: 价格，单位：元 }
        target: { $ref: '#/components/schemas/CourseTarget' }
        description:
          type: [string, 'null']
          description: 课程介绍
        createTime: { $ref: '#/components/schemas/DateTime' }
        updateTime: { $ref: '#/components/schemas/DateTime' }
    CourseBody:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 课程ID，修改时必填
        subject: { $ref: '#/components/schemas/Subject' }
        name: { type: string, description: 课程名称 }
        price: { type: integer, format: int32, description: 价格，单位：元 }
        target: { $ref: '#/components/schemas/CourseTarget' }
        description:
          type: [string, 'null']
          description: 课程介绍
      required: [subject, name, price, target]

    # ---------- 活动 ----------
...
    ClueBody:
      type: object
      description: 新增线索的请求体，`status` 与 `userId` 由服务端强制设置，传入不生效
      properties:
        phone: { type: string, description: 手机号，唯一 }
        channel: { $ref: '#/components/schemas/Channel' }
        activityId:
          anyOf:
            - type: integer
              format: int64
            - type: 'null'
          description: 关联活动的ID
        name:
          type: [string, 'null']
          description: 客户姓名
        gender:
          anyOf:
            - $ref: '#/components/schemas/Gender'
            - type: 'null'
        age:
          anyOf:
            - type: integer
              format: int32
            - type: 'null'
          description: 年龄
        wechat:
          type: [string, 'null']
          description: 微信号
        qq:
          type: [string, 'null']
          description: QQ号
        subject:
          anyOf:
            - $ref: '#/components/schemas/Subject'
            - type: 'null'
        level:
          anyOf:
            - $ref: '#/components/schemas/ClueLevel'
            - type: 'null'
        nextTime:
          anyOf:
            - $ref: '#/components/schemas/DateTime'
            - type: 'null'
      required: [phone, channel]
    ClueTrackBody:
      type: object
      description: 跟进线索的请求体，未传字段保持原值
      properties:
        id: { type: integer, format: int64, description: 线索ID }
        name:
          type: [string, 'null']
          description: 客户姓名
        phone:
          type: [string, 'null']
          description: 手机号
        channel:
          anyOf:
            - $ref: '#/components/schemas/Channel'
            - type: 'null'
        gender:
          anyOf:
            - $ref: '#/components/schemas/Gender'
            - type: 'null'
        age:
          anyOf:
            - type: integer
              format: int32
            - type: 'null'
          description: 年龄
        wechat:
          type: [string, 'null']
          description: 微信号
        qq:
          type: [string, 'null']
          description: QQ号
        subject:
          anyOf:
            - $ref: '#/components/schemas/Subject'
            - type: 'null'
        level:
          anyOf:
            - $ref: '#/components/schemas/ClueLevel'
            - type: 'null'
        nextTime:
          anyOf:
            - $ref: '#/components/schemas/DateTime'
            - type: 'null'
        record:
          type: [string, 'null']
          description: 本次跟进内容，写入跟进记录
      required: [id]
    MarkFalseClueBody:
      type: object
      properties:
        reason: { $ref: '#/components/schemas/FalseReason' }
        remark:
          type: [string, 'null']
          description: 备注信息，写入跟进记录的 `record`
      required: [reason]

    # ---------- 商机 ----------
...
    CustomerBody:
      type: object
      description: 新增或修改客户的请求体
      properties:
        id:
          type: integer
          format: int64
          description: 客户ID，修改时必填
        phone: { type: string, description: 手机号，唯一 }
        channel: { $ref: '#/components/schemas/Channel' }
        name:
          type: [string, 'null']
          description: 客户姓名
        gender:
          anyOf:
            - $ref: '#/components/schemas/Gender'
            - type: 'null'
        age:
          anyOf:
            - type: integer
              format: int32
            - type: 'null'
          description: 年龄
        wechat:
          type: [string, 'null']
          description: 微信号
        qq:
          type: [string, 'null']
          description: QQ号
        degree:
          anyOf:
            - $ref: '#/components/schemas/Degree'
            - type: 'null'
        jobStatus:
          anyOf:
            - $ref: '#/components/schemas/JobStatus'
            - type: 'null'
        subject:
          anyOf:
            - $ref: '#/components/schemas/Subject'
            - type: 'null'
        courseId:
          anyOf:
            - type: integer
              format: int64
            - type: 'null'
          description: 意向课程ID
      # channel 按页面原型（3.3 渠道来源，选填）与接口文档均为非必须
      required: [phone]

    # ---------- 操作日志 ----------
...
    CoursePage:
      type: object
      properties:
        total: { type: integer, format: int64, description: 总记录数 }
        rows:
          type: array
          items: { $ref: '#/components/schemas/CourseVO' }
      required: [total, rows]
...
    CluePage:
      type: object
      properties:
        total: { type: integer, format: int64, description: 总记录数 }
        rows:
          type: array
          items: { $ref: '#/components/schemas/Clue' }
      required: [total, rows]
...
    CustomerPage:
      type: object
      properties:
        total: { type: integer, format: int64, description: 总记录数 }
        rows:
          type: array
          items: { $ref: '#/components/schemas/Customer' }
      required: [total, rows]
...
    ResultCoursePage:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/CoursePage' }
    ResultCourseList:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/CourseVO' }
    ResultCourse:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data:
              anyOf:
                - $ref: '#/components/schemas/CourseVO'
                - type: 'null'
...
    ResultCluePage:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/CluePage' }
    ResultClue:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data:
              anyOf:
                - $ref: '#/components/schemas/Clue'
                - type: 'null'
...
    ResultCustomerPage:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/CustomerPage' }
    ResultCustomer:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data:
              anyOf:
                - $ref: '#/components/schemas/Customer'
                - type: 'null'
...